+++ apps/InstallerSkeleton.kdapp/scripts/reinstall.sh
@@ -1,9 +1,17 @@
 #       Setup - DO NOT TOUCH        #
 # --------------------------------- #
-USER=$1
-OUT=$2
-rm -rf $OUT/*
-mkdir -p $OUT
+if [ $# -lt 2 ]
+  then
+  echo "Please provide a username and log output"
+  echo "bash reinstall.sh <user> <log output>"
+  echo ""
+  exit 0
+else
+  USER=$1
+  OUT=$2
+  rm -rf $OUT/*
+  mkdir -p $OUT
+fi
 
 #       Start Coding Here...        #
 # --------------------------------- #
